src/requirements.txt
Flask
azure-cosmosdb-table
uuid
azure-keyvault-secrets
azure.identity 